% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{cv}
\alias{cv}
\title{cv}
\usage{
cv(data, kernel = "gaussian", bw)
}
\arguments{
\item{data}{numeric vector, the data from which the estimate is to be computed}

\item{kernel}{a character string, naming the type of kernel function to use. This must match one of "gaussian", "rectangular", "triangular", "epanechnikov", "biweight", "cosine", with default "gaussian"}

\item{bw}{a single number, specifying the bandwidth. There is also a function to help the user to find the appropriate bandwidth}
}
\value{
the integrated squared error estimated by cross validation with bandwidth bw
}
\description{
compute the integrated squared error estimated by cross validation
}
\examples{
\dontrun{
data <- rexp(1000,1)
cv(data, "gaussian", 0.05)
}
}
