% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gcd}
\alias{gcd}
\title{gcd}
\usage{
gcd(a, b)
}
\arguments{
\item{a}{an integer}

\item{b}{an integer}
}
\value{
x the greatest common divisor of two numbers
}
\description{
Calculate the greatest common divisor of two numbers
}
\examples{
\dontrun{
a <- 6
b <- 8
mygcd(6,8);}
}
