% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ker.bandwidth.R
\name{ker.bandwidth}
\alias{ker.bandwidth}
\title{ker.bandwidth}
\usage{
ker.bandwidth(data, rule = "Silverman")
}
\arguments{
\item{data}{numeric vector, naming the data set}

\item{rule}{a character string, naming the rule to use. This should match one of "Scott", "Silverman", with the default being "Silverman"}
}
\value{
a single number, which is just the bandwidth
}
\description{
get the bandwidth of a data set
}
\examples{
\dontrun{
library("gss")
data(buffalo)
bw <- ker.bandwidth(buffalo, "Silverman")
}
}
