% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ker.R
\name{ker}
\alias{ker}
\title{ker}
\usage{
ker(x, type = "gaussian")
}
\arguments{
\item{x}{a single number, the point to calculate}

\item{type}{a character stirng, naming the type of kernel function to use. This must match one of "gaussian", "rectangular", "triangular", "epanechnikov", "biweight", "cosine", with default "gaussian"}
}
\value{
a single number, the value of the kernel function at point x
}
\description{
compute the value of a certain type of kernel function at a certain point, with total 6 types of kernels
}
\examples{
\dontrun{
x <- ker(1,"gaussian")
y <- ker(2,"biweight")
}
}
