% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MinMerge}
\alias{MinMerge}
\title{MinMerge}
\usage{
MinMerge(x, y)
}
\arguments{
\item{x}{a vector}

\item{y}{a vector}
}
\value{
a vector consisting of the minimum of the two vectors
}
\description{
Take the minimum of two vectors and merge them
}
