% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rbivariate_c}
\alias{rbivariate_c}
\title{rbivariate_c}
\usage{
rbivariate_c(N, n, a, b)
}
\arguments{
\item{N}{the number of samples}

\item{n}{the parameter}

\item{a}{the parameter}

\item{b}{the parameter}
}
\value{
a random sample matrix of size N
}
\description{
Generate a bivariate sampler using Rcpp
}
\examples{
\dontrun{
rbivariate_c(1000, 25, 1, 1)
}
}
