\name{StatComp21085-package}
\alias{StatComp21085-package}
\docType{package}
\title{
  Bandwidth selection and kernel density estimation
}
\description{
  Using kernel density estimation to estimate the data density function. As for kernel function, this package use Gaussian, Rectangular, stationarity, Epanechnikov, biweight and cosine six basic kernel functions. In the bandwidth selection part, Scott rule and Silverman rule are given. There is also a function that calculates the integrated squared error estimated by cross validation when given bandwidth, so we can get the best bandwidth through cross validation.
}
\author{
Xiaoke Zhang

Maintainer: Xiaoke Zhang <zxk170091@mail.ustc.edu.cn>
}
\references{
  Härdle, Wolfgang Karl, et al. Nonparametric and semiparametric models. Springer Science & Business Media, 2004.
}
\keyword{ package }
